+++ waterfun-service/src/main/java/org/waterwood/waterfunservice/api/response/post/PostCardResp.java
@@ -27,8 +27,8 @@ public class PostCardResp {
     private UserBrief userBrief;
 
     private CloudResPresignedUrlResp coverImage;
-    private OptionVO<Integer> category;
-    private List<OptionVO<Integer>> tags;
+    private OptionVO<Long> category;
+    private List<OptionVO<Long>> tags;
 
     private Long viewCount;
     private Long likeCount;
